Add config run to allow for running all shakes from the config

+++ shake.core/Utils/ConfigModels/AudioShake.cs
@@ -4,6 +4,6 @@ public class AudioShake
 {
     public string Id { get; set; }
     public Guid AudioDevice { get; set; }
-    public long Delay { get; set; }
+    public int Delay { get; set; }
     public bool UseTestAudio { get; set; }
 }
