+++ shake.core/Utils/ConfigModels/VideoShake.cs
@@ -2,6 +2,6 @@
 
 public class VideoShake
 {
-    public long Timer { get; set; }
+    public int Timer { get; set; }
     public bool IncludeDisplay { get; set; }
 }
